pyproject.toml
[tool.poetry]
name = "dashboard-llm"
version = "0.1.0"
description = ""
authors = ["Vejeja <kornilova.valeriia@gmail.com>"]

packages = [
  { include = "monitoring_dashboard", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
sentence-transformers = "^2.2.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[tool.poetry.scripts]
monitoring-dashboard = "main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
